It is not a good idea to try to speed up exp()
I don't see significant speedup even in the best of cases (25% being the
best in one set) and sometimes the fastexp() method is slower.  I think
the difference in results has to do with the load on the machine.  But I
don't like the idea of something that might be slower, is never much
faster, and fails to give answers identical to what we had before in oce
(and what is documented in the various papers on the Barnes method).